[BOJ] 15654.N과 M(5)
